Remove card folding from Redux (#112)
Card folding state is typically something that is done with local state. That's way easier than using Redux. It is also a first prep for React 16.7, which will introduce hooks to do this in an even more simple way. That version will most likely also introduce functionality to fetch data in an easier manner than what is currently done with Redux. However, I think it is better for now to leave that part, as that is not much cleaner using local state.